+++ spark/v3.4/spark/src/main/java/org/apache/iceberg/spark/source/metrics/ScannedDataManifests.java
@@ -31,7 +31,7 @@ public class ScannedDataManifests extends CustomSumMetric {
 
   @Override
   public String description() {
-    return "num scanned data manifests";
+    return "number of scanned data manifests";
   }
 
   @Override
