+++ spark/v3.4/spark/src/main/java/org/apache/iceberg/spark/source/metrics/scannedDataFiles.java
@@ -18,30 +18,19 @@
  */
 package org.apache.iceberg.spark.source.metrics;
 
-import org.apache.iceberg.metrics.CounterResult;
-import org.apache.iceberg.metrics.ScanReport;
-import org.apache.spark.sql.connector.metric.CustomTaskMetric;
+import org.apache.spark.sql.connector.metric.CustomSumMetric;
 
-public class TaskResultDataFiles implements CustomTaskMetric {
-  private final long value;
+public class scannedDataFiles extends CustomSumMetric {
 
-  private TaskResultDataFiles(long value) {
-    this.value = value;
-  }
+  static final String NAME = "scannedDataFiles";
 
   @Override
   public String name() {
-    return ResultDataFiles.NAME;
+    return NAME;
   }
 
   @Override
-  public long value() {
-    return value;
-  }
-
-  public static TaskResultDataFiles from(ScanReport scanReport) {
-    CounterResult counter = scanReport.scanMetrics().resultDataFiles();
-    long value = counter != null ? counter.value() : -1;
-    return new TaskResultDataFiles(value);
+  public String description() {
+    return "number of scanned data files";
   }
 }
